CenterServer/requirements.txt
-i https://pypi.org/simple
aioredis==1.3.1
amqp==5.1.1 ; python_version >= '3.6'
asgiref==3.5.2 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.1.0 ; python_version >= '3.5'
autobahn==22.7.1 ; python_version >= '3.7'
automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24 ; python_version >= '3.6'
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1 ; python_version >= '3.6'
daphne==3.0.2 ; python_version >= '3.6'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.24.2
drf-yasg==1.20.3
gunicorn==20.1.0
hiredis==2.0.0 ; python_version >= '3.6'
hyperlink==21.0.0
idna==3.4 ; python_version >= '3.5'
incremental==21.3.0
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.16.0 ; python_version >= '3.7'
kombu==5.2.4 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
msgpack==1.0.4
oauthlib==3.2.1 ; python_version >= '3.6'
packaging==21.3
pillow==9.2.0
prompt-toolkit==3.0.31 ; python_full_version >= '3.6.2'
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygments==2.13.0
pyjwt==2.5.0 ; python_version >= '3.7'
pyopenssl==22.1.0
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pyrsistent==0.18.1 ; python_version >= '3.7'
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.1
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ruamel.yaml==0.17.21 ; python_version >= '3'
ruamel.yaml.clib==0.2.6 ; python_version < '3.11' and platform_python_implementation == 'CPython'
service-identity==21.1.0
setuptools==65.4.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==4.0.0
social-auth-core==4.3.0 ; python_version >= '3.6'
sqlparse==0.4.3 ; python_version >= '3.5'
twisted[tls]==22.8.0 ; python_full_version >= '3.7.1'
txaio==22.2.1 ; python_version >= '3.6'
typing-extensions==4.3.0 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.12
vine==5.0.0 ; python_version >= '3.6'
wcwidth==0.2.5
zope.interface==5.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pandas
numpy
scikit-learn
graphviz
beautifulsoup4
APScheduler
twilio
xlsxwriter
aspose-words==23.2.0
python-docx==0.8.11
reportlab
elasticsearch
redis
paramiko